Conflicts:
	protected/migrations/m140208_146057_exemption.php

+++ protected/migrations/m140408_180000_create_study_year_table.php
@@ -2,18 +2,21 @@
 
 class m140408_180000_create_study_year_table extends CDbMigration
 {
-	public function up()
-	{
-        $this->createTable('study_year',array(
+    public function up()
+    {
+        if (Yii::app()->getDb()->schema->getTable('study_year') !== null) {
+            $this->dropTable('study_year');
+        }
+        $this->createTable('study_year', array(
             'id' => 'pk',
             'title' => 'varchar(10) NOT NULL',
         ));
-	}
+    }
 
-	public function down()
-	{
-		$this->dropTable('study_year');
-		return true;
-	}
+    public function down()
+    {
+        $this->dropTable('study_year');
+        return true;
+    }
 
 }
